requirements.txt
django-environ==0.10.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
argon2-cffi==23.1.0
djoser==2.2.0 
djangorestframework-simplejwt==5.2.2 
PyJWT==2.8.0
social-auth-core==4.4.2
social-auth-app-django==5.2.0   
Pillow==10.0.0
django-storages==1.13.2
django-ckeditor==6.7.0
psycopg2==2.9.7
python-slugify==8.0.1
requests==2.31.0
whitenoise==6.5.0
gunicorn==21.2.0  